1. Added logic to monitor valve opening so that the display and serial output keep going during this process.
2. Changed to just monitor the valve-open signal since with this particular valve, we never want to
close it all the way (because it over-closes -- ug).

3. Added logic to operational part to be careful to not change the valve or heater relay or solar pump
to often (not more than once every 15 minutes).  THis is just in case of a defective temperature sensor
or circuit that might otherwise cause these mechanical things to beat themselves to death.

This is pretty close to ready for initial testing in situ.f